refactor: move product related controller to product.controller
