+++ finpro-api/src/controllers/product.controller/product.controller.ts
@@ -1,32 +1,21 @@
-import { prisma } from '../prisma';
-import { ICloudinaryResult, IProduct, IProductImage } from '../types/product.type';
-import { NextFunction, Request, Response } from 'express';
-import {v2 as cloudinary} from 'cloudinary';
-import { Prisma } from '@prisma/client';
-
+import { prisma } from "../../prisma";
+import { ICloudinaryResult, IProduct, IProductImage } from "../../types/product.type";
+import { NextFunction, Request, Response } from "express";
+import { v2 as cloudinary } from "cloudinary";
+import { Prisma } from "@prisma/client";
 
 export class ProductController {
   async createProduct(req: Request, res: Response, next: NextFunction) {
     try {
-      const {
-        name,
-        price,
-        productSubCategoryId,
-        brandId,
-        description,
-        sku,
-        barcode,
-        weight,
-        dimensions,
-        images
-      } = req.body;
-      
-      const slug = name.toLowerCase().replace(/\s+/g, '-');
+      const { name, price, productSubCategoryId, brandId, description, sku, barcode, weight, dimensions, images } =
+        req.body;
+
+      const slug = name.toLowerCase().replace(/\s+/g, "-");
 
       if (!name || !productSubCategoryId || !price || images.length === 0) {
         res.status(400).json({
           success: false,
-          message: 'Name, Price, Product Sub Category, and Product Images are required',
+          message: "Name, Price, Product Sub Category, and Product Images are required",
         });
       }
 
@@ -44,8 +33,8 @@ export class ProductController {
             weight,
             dimensions,
           },
-        })
-          
+        });
+
         const productId = product.id;
 
         const productImages = await tx.productImage.createMany({
@@ -54,15 +43,15 @@ export class ProductController {
             imageUrl: image.secure_url,
             cldPublicId: image.public_id,
             isMainImage: image.isMainImage,
-          }))
-        })
-        return {product, productImages};
-      })
+          })),
+        });
+        return { product, productImages };
+      });
 
       res.status(201).json({
         success: true,
-        message: 'Product created successfully',
-        newProduct
+        message: "Product created successfully",
+        newProduct,
       });
     } catch (error) {
       next(error);
@@ -78,10 +67,7 @@ export class ProductController {
         secure: true,
       });
       const { paramsToSign } = req.body;
-      const signature = cloudinary.utils.api_sign_request(
-        paramsToSign,
-        process.env.CLOUDINARY_API_SECRET as string
-      )
+      const signature = cloudinary.utils.api_sign_request(paramsToSign, process.env.CLOUDINARY_API_SECRET as string);
       res.status(200).json({
         signature,
       });
@@ -97,16 +83,16 @@ export class ProductController {
           productImage: true,
           productSubCategory: {
             include: {
-              productCategory: true
-            }
+              productCategory: true,
+            },
           },
           productBrand: true,
-        }
+        },
       });
       res.status(200).json({
         success: true,
-        message: 'Products fetched successfully',
-        products
+        message: "Products fetched successfully",
+        products,
       });
     } catch (error) {
       next(error);
@@ -124,16 +110,16 @@ export class ProductController {
           productImage: true,
           productSubCategory: {
             include: {
-              productCategory: true
-            }
+              productCategory: true,
+            },
           },
           productBrand: true,
-        }
+        },
       });
       res.status(200).json({
         success: true,
-        message: 'Product fetched successfully',
-        product
+        message: "Product fetched successfully",
+        product,
       });
     } catch (error) {
       next(error);
